Refactor project structure and update README
- Move Pydantic models to a dedicated `src/models` directory.
- Update import statements to reflect the new model location.
- Overhaul the `README.md` to provide accurate and comprehensive
  documentation for the project.
- The new README includes an updated project structure, detailed
  setup and usage instructions, and an example output.
